requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2

# Data Processing
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0

# Web Scraping & APIs
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.25.1
aiohttp==3.9.1

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.16.3

# AI/ML
openai==1.3.7
scikit-learn==1.3.2

# Dashboard & Visualization
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Email & Notifications
sendgrid==6.11.0

# Report Generation
reportlab==4.0.7
openpyxl==3.1.2

# Async Support
asyncio==3.4.3

# Data Validation
phonenumbers==8.13.26
email-validator==2.1.0
us==3.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
faker==21.0.0